changed render city info to render an array of cards and changed the info to display as a card
