+++ src/components/pages/CityInfo/RenderCityInfo.js
@@ -1,5 +1,6 @@
 import React, { useState } from 'react';
 import { Layout, Card } from 'antd';
+import CityCard from './CityCard';
 import Title from 'antd/lib/typography/Title';
 import MapContainer from '../Home/Map';
 const mapboxgl = require('mapbox-gl');
@@ -9,48 +10,6 @@ const { Header } = Layout;
 
 function RenderCityInfo(props) {
   const { city } = props;
-  const [tab, setTab] = useState('general');
-
-  const tabChange = key => {
-    setTab(key);
-  };
-  const tabList = [
-    {
-      key: 'general',
-      tab: 'general',
-    },
-    {
-      key: 'living',
-      tab: 'living',
-    },
-    {
-      key: 'crime',
-      tab: 'crime',
-    },
-  ];
-  const contentList = {
-    general: (
-      <div>
-        <h1>{city.State}</h1>
-        <p>Population: {city.Population}</p>
-        <p>Density: {city.Density}</p>
-        <p>ZIP code: {city.ZIPcode}</p>
-      </div>
-    ),
-    living: (
-      <div>
-        <p>Cost of Living Index: {city.CostOfLivingIndex}</p>
-        <p>Walkability Score: {city.WalkScore}</p>
-        <p>Walkability: {city.WalkScoreDescription}</p>
-      </div>
-    ),
-    crime: (
-      <div>
-        <p>Property Crime: {city.PropertyCrimeRate}</p>
-        <p>Violent Crime: {city.ViolentCrimeRate}</p>
-      </div>
-    ),
-  };
 
   return (
     <div>
@@ -63,17 +22,9 @@ function RenderCityInfo(props) {
           </Header>
         </div>
         <div>
-          <Card
-            style={{ width: 500 }}
-            title={city.City}
-            tabList={tabList}
-            activeTabKey={tab}
-            onTabChange={tab => {
-              tabChange(tab);
-            }}
-          >
-            {contentList[tab]}
-          </Card>
+          {city.map(city => (
+            <CityCard key={city.id} city={city} />
+          ))}
         </div>
       </Layout>
     </div>
